Moved Token and TokenRepository into appropriate packages.
