+++ src/main/java/uk/co/dashery/controller/TokenController.java
@@ -2,7 +2,7 @@ package uk.co.dashery.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
-import uk.co.dashery.Token;
+import uk.co.dashery.data.Token;
 import uk.co.dashery.service.TokenService;
 
 import java.util.List;
