+++ src/main/java/uk/co/dashery/repository/TokenRepository.java
@@ -1,6 +1,7 @@
-package uk.co.dashery;
+package uk.co.dashery.repository;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
+import uk.co.dashery.data.Token;
 
 public interface TokenRepository extends MongoRepository<Token, String> {
