+++ src/test/java/uk/co/dashery/service/TokenServiceTest.java
@@ -4,8 +4,8 @@ import org.junit.Before;
 import org.junit.Test;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import uk.co.dashery.Token;
-import uk.co.dashery.TokenRepository;
+import uk.co.dashery.data.Token;
+import uk.co.dashery.repository.TokenRepository;
 
 import java.util.List;
